Conflicts:
	src/main/java/de/htw/sdf/photoplatform/webservice/controller/PurchaseController.java
	src/main/resources/messages_de.properties
	src/main/webapp/index.html

+++ src/main/java/de/htw/sdf/photoplatform/manager/PurchaseManager.java
@@ -9,6 +9,7 @@ import de.htw.sdf.photoplatform.persistence.model.Image;
 import de.htw.sdf.photoplatform.persistence.model.PurchaseItem;
 import de.htw.sdf.photoplatform.persistence.model.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -96,7 +97,7 @@ public interface PurchaseManager {
      * @param images list of images to buy.
      * @return total price of all images.
      */
-    Double calculatePrice(List<PurchaseItem> images);
+    BigDecimal calculatePrice(List<PurchaseItem> images);
 
 
     /**
